convert to TS

+++ packages/cli/test/integration.test.ts
@@ -1,5 +1,3 @@
-// @ts-nocheck
-// Note: this file is incrementally migrating to typescript
 import ms from 'ms';
 import path from 'path';
 import { URL, parse as parseUrl } from 'url';
@@ -8,7 +6,8 @@ import { Readable } from 'stream';
 import { homedir, tmpdir } from 'os';
 import _execa from 'execa';
 import XDGAppPaths from 'xdg-app-paths';
-import fetch from 'node-fetch';
+import fetch, { RequestInfo } from 'node-fetch';
+// @ts-ignore
 import tmp from 'tmp-promise';
 import retry from 'async-retry';
 import fs, {
@@ -25,17 +24,48 @@ import pkg from '../package.json';
 import prepareFixtures from './helpers/prepare';
 import { fetchTokenWithRetry } from '../../../test/lib/deployment/now-deploy';
 import { once } from 'node:events';
+import http from 'http';
 
 const TEST_TIMEOUT = 3 * 60 * 1000;
 jest.setTimeout(TEST_TIMEOUT);
 
-// log command when running `execa`
-function execa(file, args, options) {
-  console.log(`$ vercel ${args.join(' ')}`);
-  return _execa(file, args, options);
+type BoundChildProcess = _execa.ExecaChildProcess & {
+  stdout: Readable;
+  stdin: Readable;
+  stderr: Readable;
+};
+
+interface TmpDir {
+  name: string;
+  removeCallback: () => void;
 }
 
-function fixture(name) {
+interface Build {
+  use: string;
+}
+
+// log command when running `execa`
+function execa(
+  file: string,
+  args: string[],
+  options?: _execa.Options<string>
+): BoundChildProcess {
+  console.log(`$ vercel ${args.join(' ')}`);
+  const proc = _execa(file, args, options);
+  if (proc.stdin === null) {
+    throw new Error(`vercel ${args.join(' ')} - failed to bind to stdin`);
+  }
+  if (proc.stdout === null) {
+    throw new Error(`vercel ${args.join(' ')} - failed to bind to stdout`);
+  }
+  if (proc.stderr === null) {
+    throw new Error(`vercel ${args.join(' ')} - failed to bind to stderr`);
+  }
+  // Typescript should know this can match because of the guards above
+  return proc as BoundChildProcess;
+}
+
+function fixture(name: string) {
   const directory = path.join(tmpFixturesDir, name);
   const config = path.join(directory, 'project.json');
 
@@ -48,21 +78,21 @@ function fixture(name) {
 }
 
 const binaryPath = path.resolve(__dirname, `../scripts/start.js`);
-const example = name =>
+const example = (name: string) =>
   path.join(__dirname, '..', '..', '..', 'examples', name);
 const deployHelpMessage = `${logo} vercel [options] <command | path>`;
 let session = 'temp-session';
 
 const isCanary = pkg.version.includes('canary');
 
-const pickUrl = stdout => {
+const pickUrl = (stdout: string) => {
   const lines = stdout.split('\n');
   return lines[lines.length - 1];
 };
 
-const createFile = dest => fs.closeSync(fs.openSync(dest, 'w'));
+const createFile = (dest: fs.PathLike) => fs.closeSync(fs.openSync(dest, 'w'));
 
-const waitForDeployment = async href => {
+const waitForDeployment = async (href: RequestInfo) => {
   console.log(`waiting for ${href} to become ready...`);
   const start = Date.now();
   const max = ms('4m');
@@ -89,7 +119,7 @@ const waitForDeployment = async href => {
   }
 };
 
-function fetchTokenInformation(token, retries = 3) {
+function fetchTokenInformation(token: any, retries = 3) {
   const url = `https://api.vercel.com/v2/user`;
   const headers = { Authorization: `Bearer ${token}` };
 
@@ -111,7 +141,13 @@ function fetchTokenInformation(token, retries = 3) {
   );
 }
 
-function formatOutput({ stderr, stdout }) {
+function formatOutput({
+  stderr,
+  stdout,
+}: {
+  stderr: string | Readable;
+  stdout: string | Readable;
+}) {
   return `
 -----
 
@@ -127,7 +163,7 @@ ${stdout || '(no output)'}
   `;
 }
 
-async function vcLink(projectPath) {
+async function vcLink(projectPath: string) {
   const { exitCode, stdout, stderr } = await execa(
     binaryPath,
     ['link', '--yes', ...defaultArgs],
@@ -140,28 +176,57 @@ async function vcLink(projectPath) {
   expect(exitCode, formatOutput({ stdout, stderr })).toBe(0);
 }
 
+async function getLocalhost(vc: BoundChildProcess): Promise<RegExpExecArray> {
+  let localhost: RegExpExecArray | undefined;
+  await waitForPrompt(vc, chunk => {
+    if (chunk.includes('Ready! Available at')) {
+      localhost = /(https?:[^\s]+)/g.exec(chunk) || undefined;
+      return true;
+    }
+    return false;
+  });
+
+  // This should never happen because waitForPrompt will time out
+  // and never return here in this case, but extra checking is fine
+  // and it makes typescript happy
+  if (!localhost) {
+    throw new Error('Localhost not found!');
+  }
+
+  return localhost;
+}
+
+function getTmpDir(): TmpDir {
+  return tmp.dirSync({
+    // This ensures the directory gets
+    // deleted even if it has contents
+    unsafeCleanup: true,
+  }) as TmpDir;
+}
+
 // AVA's `t.context` can only be set before the tests,
 // but we want to set it within as well
-const context = {};
+const context: {
+  deployment: string | undefined;
+  secretName: string | undefined;
+} = {
+  deployment: undefined,
+  secretName: undefined,
+};
 
 const defaultOptions = { reject: false };
-const defaultArgs = [];
-let token;
-let email;
-let contextName;
+const defaultArgs: string[] = [];
+let token: string | undefined;
+let email: string | undefined;
+let contextName: string | undefined;
 
-let tmpDir;
+let tmpDir: TmpDir | undefined;
 let tmpFixturesDir = path.join(tmpdir(), 'tmp-fixtures');
 
 let globalDir = XDGAppPaths('com.vercel.cli').dataDirs()[0];
 
 if (!process.env.CI) {
-  tmpDir = tmp.dirSync({
-    // This ensures the directory gets
-    // deleted even if it has contents
-    unsafeCleanup: true,
-  });
-
+  tmpDir = getTmpDir();
   globalDir = path.join(tmpDir.name, 'com.vercel.tests');
 
   defaultArgs.push('-Q', globalDir);
@@ -171,7 +236,7 @@ if (!process.env.CI) {
   );
 }
 
-function mockLoginApi(req, res) {
+function mockLoginApi(req: http.IncomingMessage, res: http.ServerResponse) {
   const { url = '/', method } = req;
   let { pathname = '/', query = {} } = parseUrl(url, true);
   console.log(`[mock-login-server] ${method} ${pathname}`);
@@ -204,13 +269,13 @@ const loginApiServer = require('http')
     console.log(`[mock-login-server] Listening on ${loginApiUrl}`);
   });
 
-const execute = (args, options) =>
+const execute = (args: string[], options?: _execa.Options<string>) =>
   execa(binaryPath, [...defaultArgs, ...args], {
     ...defaultOptions,
     ...options,
   });
 
-const apiFetch = (url, { headers, ...options } = {}) => {
+const apiFetch = (url: string, { headers, ...options }: any = {}) => {
   return fetch(`https://api.vercel.com${url}`, {
     headers: {
       Authorization: `Bearer ${token}`,
@@ -220,14 +285,17 @@ const apiFetch = (url, { headers, ...options } = {}) => {
   });
 };
 
-const waitForPrompt = (cp, assertion) =>
-  new Promise((resolve, reject) => {
+const waitForPrompt = (
+  cp: BoundChildProcess,
+  assertion: (chunk: string) => boolean
+) =>
+  new Promise<void>((resolve, reject) => {
     console.log('Waiting for prompt...');
     const handleTimeout = setTimeout(
       () => reject(new Error('timeout in waitForPrompt')),
       TEST_TIMEOUT / 2
     );
-    const listener = chunk => {
+    const listener = (chunk: string) => {
       console.log('> ' + chunk);
       if (assertion(chunk)) {
         cp.stdout.off && cp.stdout.off('data', listener);
@@ -267,7 +335,11 @@ const createUser = async () => {
 
 const getConfigAuthPath = () => path.join(globalDir, 'auth.json');
 
-async function setupProject(process, projectName, overrides) {
+async function setupProject(
+  process: BoundChildProcess,
+  projectName: string,
+  overrides: { devCommand?: any; buildCommand?: any; outputDirectory?: any }
+) {
   await waitForPrompt(process, chunk => /Set up [^?]+\?/.test(chunk));
   process.stdin.write('yes\n');
 
@@ -387,6 +459,10 @@ test('default command should prompt login with empty auth.json', async () => {
 test(
   'login',
   async () => {
+    if (!email) {
+      throw new Error('Shared state "email" not set.');
+    }
+
     await fs.remove(getConfigAuthPath());
     const loginOutput = await execa(binaryPath, [
       'login',
@@ -1049,15 +1125,7 @@ test('Deploy `api-env` fixture and test `vercel env` command', async () => {
       cwd: target,
     });
 
-    let localhost = undefined;
-    await waitForPrompt(vc, chunk => {
-      if (chunk.includes('Ready! Available at')) {
-        localhost = /(https?:[^\s]+)/g.exec(chunk);
-        return true;
-      }
-      return false;
-    });
-
+    const localhost = await getLocalhost(vc);
     const apiUrl = `${localhost[0]}/api/get-env`;
     const apiRes = await fetch(apiUrl);
 
@@ -1087,15 +1155,7 @@ test('Deploy `api-env` fixture and test `vercel env` command', async () => {
       cwd: target,
     });
 
-    let localhost = undefined;
-    await waitForPrompt(vc, chunk => {
-      if (chunk.includes('Ready! Available at')) {
-        localhost = /(https?:[^\s]+)/g.exec(chunk);
-        return true;
-      }
-      return false;
-    });
-
+    const localhost = await getLocalhost(vc);
     const apiUrl = `${localhost[0]}/api/get-env`;
     const apiRes = await fetch(apiUrl);
     expect(apiRes.status).toBe(200);
@@ -1167,15 +1227,7 @@ test('Deploy `api-env` fixture and test `vercel env` command', async () => {
       cwd: target,
     });
 
-    let localhost = undefined;
-    await waitForPrompt(vc, chunk => {
-      if (chunk.includes('Ready! Available at')) {
-        localhost = /(https?:[^\s]+)/g.exec(chunk);
-        return true;
-      }
-      return false;
-    });
-
+    const localhost = await getLocalhost(vc);
     const apiUrl = `${localhost[0]}/api/get-env`;
     const apiRes = await fetch(apiUrl);
 
@@ -1262,7 +1314,7 @@ test('Deploy `api-env` fixture and test `vercel env` command', async () => {
     expect(exitCode, formatOutput({ stdout, stderr })).toBe(0);
   }
 
-  function vcEnvRemoveByName(name) {
+  function vcEnvRemoveByName(name: string) {
     return execa(binaryPath, ['env', 'rm', name, '-y', ...defaultArgs], {
       reject: false,
       cwd: target,
@@ -1717,6 +1769,10 @@ test('ensure we render a warning for deployments with no files', async () => {
 });
 
 test('output logs with "short" output', async () => {
+  if (!context.deployment) {
+    throw new Error('Shared state "context.deployment" not set.');
+  }
+
   const { stderr, stdout, exitCode } = await execa(
     binaryPath,
     ['logs', context.deployment, ...defaultArgs],
@@ -1738,6 +1794,10 @@ test('output logs with "short" output', async () => {
 });
 
 test('output logs with "raw" output', async () => {
+  if (!context.deployment) {
+    throw new Error('Shared state "context.deployment" not set.');
+  }
+
   const { stderr, stdout, exitCode } = await execa(
     binaryPath,
     ['logs', context.deployment, ...defaultArgs, '--output', 'raw'],
@@ -2056,7 +2116,7 @@ test('try to deploy with non-existing team', async () => {
 });
 
 test('initialize example "angular"', async () => {
-  tmpDir = tmp.dirSync({ unsafeCleanup: true });
+  tmpDir = getTmpDir();
   const cwd = tmpDir.name;
   const goal = '> Success! Initialized "angular" example in';
 
@@ -2082,7 +2142,7 @@ test('initialize example "angular"', async () => {
 });
 
 test('initialize example ("angular") to specified directory', async () => {
-  tmpDir = tmp.dirSync({ unsafeCleanup: true });
+  tmpDir = getTmpDir();
   const cwd = tmpDir.name;
   const goal = '> Success! Initialized "angular" example in';
 
@@ -2111,7 +2171,7 @@ test('initialize example ("angular") to specified directory', async () => {
 });
 
 test('initialize example to existing directory with "-f"', async () => {
-  tmpDir = tmp.dirSync({ unsafeCleanup: true });
+  tmpDir = getTmpDir();
   const cwd = tmpDir.name;
   const goal = '> Success! Initialized "angular" example in';
 
@@ -2142,7 +2202,7 @@ test('initialize example to existing directory with "-f"', async () => {
 });
 
 test('try to initialize example to existing directory', async () => {
-  tmpDir = tmp.dirSync({ unsafeCleanup: true });
+  tmpDir = getTmpDir();
   const cwd = tmpDir.name;
   const goal =
     'Error: Destination path "angular" already exists and is not an empty directory. You may use `--force` or `-f` to override it.';
@@ -2159,7 +2219,7 @@ test('try to initialize example to existing directory', async () => {
 });
 
 test('try to initialize misspelled example (noce) in non-tty', async () => {
-  tmpDir = tmp.dirSync({ unsafeCleanup: true });
+  tmpDir = getTmpDir();
   const cwd = tmpDir.name;
   const goal =
     'Error: No example found for noce, run `vercel init` to see the list of available examples.';
@@ -2171,7 +2231,7 @@ test('try to initialize misspelled example (noce) in non-tty', async () => {
 });
 
 test('try to initialize example "example-404"', async () => {
-  tmpDir = tmp.dirSync({ unsafeCleanup: true });
+  tmpDir = getTmpDir();
   const cwd = tmpDir.name;
   const goal =
     'Error: No example found for example-404, run `vercel init` to see the list of available examples.';
@@ -2189,7 +2249,7 @@ test('try to revert a deployment and assign the automatic aliases', async () => 
   const secondDeployment = fixture('now-revert-alias-2');
 
   const { name } = JSON.parse(
-    fs.readFileSync(path.join(firstDeployment, 'now.json'))
+    fs.readFileSync(path.join(firstDeployment, 'now.json')).toString()
   );
   expect(name).toBeTruthy();
 
@@ -2301,7 +2361,7 @@ test('`vercel rm` removes a deployment', async () => {
         session,
         ...defaultArgs,
         '-V',
-        2,
+        '2',
         '--force',
         '--yes',
       ],
@@ -2360,7 +2420,7 @@ test('render build errors', async () => {
 });
 
 test('invalid deployment, projects and alias names', async () => {
-  const check = async (...args) => {
+  const check = async (...args: string[]) => {
     const output = await execute(args);
     expect(output.exitCode, formatOutput(output)).toBe(1);
     expect(output.stderr).toMatch(/The provided argument/gm);
@@ -2415,11 +2475,11 @@ test('create zero-config deployment', async () => {
 
   expect(data.error).toBe(undefined);
 
-  const validBuilders = data.builds.every(build =>
+  const validBuilders = data.builds.every((build: Build) =>
     isCanary ? build.use.endsWith('@canary') : !build.use.endsWith('@canary')
   );
 
-  const buildList = JSON.stringify(data.builds.map(b => b.use));
+  const buildList = JSON.stringify(data.builds.map((b: Build) => b.use));
   const message = `builders match canary (${isCanary}): ${buildList}`;
   expect(validBuilders, message).toBe(true);
 });
@@ -2449,7 +2509,6 @@ test('vercel secret ls', async () => {
   const output = await execute(['secret', 'ls']);
   expect(output.exitCode, formatOutput(output)).toBe(0);
   expect(output.stdout).toMatch(/Secrets found under/gm);
-  expect(output.stdout).toMatch(new RegExp());
 });
 
 test('vercel secret ls --test-warning', async () => {
@@ -2462,6 +2521,10 @@ test('vercel secret ls --test-warning', async () => {
 });
 
 test('vercel secret rename', async () => {
+  if (!context.secretName) {
+    throw new Error('Shared state "context.secretName" not set.');
+  }
+
   const nextName = `renamed-secret-${Date.now().toString(36)}`;
   const output = await execute([
     'secret',
@@ -2475,6 +2538,10 @@ test('vercel secret rename', async () => {
 });
 
 test('vercel secret rm', async () => {
+  if (!context.secretName) {
+    throw new Error('Shared state "context.secretName" not set.');
+  }
+
   const output = await execute(['secret', 'rm', context.secretName, '-y']);
   expect(output.exitCode, formatOutput(output)).toBe(0);
 });
@@ -2574,6 +2641,10 @@ test('fail to add a domain without a project', async () => {
 test(
   'change user',
   async () => {
+    if (!email) {
+      throw new Error('Shared state "email" not set.');
+    }
+
     const { stdout: prevUser } = await execute(['whoami']);
 
     // Delete the current token
@@ -2673,13 +2744,19 @@ test('should show prompts to set up project during first deploy', async () => {
   // and output directory
   let stderr = '';
   const port = 58351;
-  const dev = execa(binaryPath, ['dev', '--listen', port, dir, ...defaultArgs]);
+  const dev = execa(binaryPath, [
+    'dev',
+    '--listen',
+    port.toString(),
+    dir,
+    ...defaultArgs,
+  ]);
   dev.stderr.setEncoding('utf8');
 
   try {
     dev.stdout.pipe(process.stdout);
     dev.stderr.pipe(process.stderr);
-    await new Promise((resolve, reject) => {
+    await new Promise<void>((resolve, reject) => {
       dev.once('close', (code, signal) => {
         reject(`"vc dev" failed with ${signal || code}`);
       });
@@ -3143,7 +3220,7 @@ test(
     const packageJsonOriginal = await readFile(packageJsonPath, 'utf8');
     const pkg = JSON.parse(packageJsonOriginal);
 
-    async function tryDeploy(cwd) {
+    async function tryDeploy(cwd: string) {
       const { exitCode, stdout, stderr } = await execa(
         binaryPath,
         [...defaultArgs, '--public', '--yes'],
@@ -3356,12 +3433,16 @@ test('[vc dev] should show prompts to set up project', async () => {
   // remove previously linked project if it exists
   await remove(path.join(dir, '.vercel'));
 
-  const dev = execa(binaryPath, ['dev', '--listen', port, ...defaultArgs], {
-    cwd: dir,
-    env: {
-      FORCE_TTY: '1',
-    },
-  });
+  const dev = execa(
+    binaryPath,
+    ['dev', '--listen', port.toString(), ...defaultArgs],
+    {
+      cwd: dir,
+      env: {
+        FORCE_TTY: '1',
+      },
+    }
+  );
 
   await setupProject(dev, projectName, {
     buildCommand: `mkdir -p o && echo '<h1>custom hello</h1>' > o/index.html`,
@@ -3463,12 +3544,16 @@ test('[vc dev] should send the platform proxy request headers to frontend dev se
   // remove previously linked project if it exists
   await remove(path.join(dir, '.vercel'));
 
-  const dev = execa(binaryPath, ['dev', '--listen', port, ...defaultArgs], {
-    cwd: dir,
-    env: {
-      FORCE_TTY: '1',
-    },
-  });
+  const dev = execa(
+    binaryPath,
+    ['dev', '--listen', port.toString(), ...defaultArgs],
+    {
+      cwd: dir,
+      env: {
+        FORCE_TTY: '1',
+      },
+    }
+  );
 
   await setupProject(dev, projectName, {
     buildCommand: `mkdir -p o && echo '<h1>custom hello</h1>' > o/index.html`,
